Separate Sovler
We'll need this to be separate so that the API can preview a given
solution for UX goodness.
